+++ src/docs/Select/select.js
@@ -1,14 +1,14 @@
 // import { select } from '../../js/components/select.js'
-import { Select } from '../../js/components/select_new.js'
+import { select } from '../../js/components/select.js'
 
 export function selectExample(context) {
   if (!context) {
     return
   }
 
-  context.addEventListener('change', (event) => console.log('change!', [...event.target.selectedOptions].map((option) => option.value)))
+  // context.addEventListener('change', (event) => console.log('change!', [...event.target.selectedOptions].map((option) => option.value)))
 
   for (const el of context.querySelectorAll('.select')) {
-    new Select(el)
+    select(el)
   }
 }
