+++ src/docs/Select/single.html
@@ -19,7 +19,7 @@
         <form class="form-group">
             <div class="select">
                 <label class="select-label" for="select-hidden-label">City</label>
-                <select class="select-options" name="timezone" id="select-hidden-label">
+                <select class="select-options" name="timezone" id="select-long">
                     <option value="" disabled selected hidden>Select City</option>
                     <option>Gävle</option>
                     <option>Göteborg</option>
