+++ src/js/components/select.js
@@ -1,273 +1,433 @@
 import { h, Ref, uid } from '../utils'
 
-function moveFocusTo(position, list) {
-  if (typeof position === 'object') {
-    if (!position.classList.contains('select-option--focus')) {
-      list.querySelector('.select-option--focus').classList.remove('select-option--focus')
-      position.classList.add('select-option--focus')
-    }
-  }
-  if (position === 'selected') {
-    const selectedItem = list.querySelector('.select-option--selected') || list.querySelector('.select-option:first-child')
-    selectedItem.classList.add('select-option--focus')
-  }
-
-  if (position === 'next') {
-    const focusedItem = list.querySelector('.select-option--focus')
-    if (!focusedItem.nextElementSibling) {
-      list.firstElementChild.classList.add('select-option--focus')
-    } else {
-      focusedItem.nextElementSibling.classList.add('select-option--focus')
-    }
-
-    focusedItem.classList.remove('select-option--focus')
-  }
-  if (position === 'prev') {
-    const focusedItem = list.querySelector('.select-option--focus')
-    if (!focusedItem.previousElementSibling) {
-      list.lastElementChild.classList.add('select-option--focus')
-    } else {
-      focusedItem.previousElementSibling.classList.add('select-option--focus')
-    }
-    focusedItem.classList.remove('select-option--focus')
-  }
-  if (position === 'first') {
-    const focusedItem = list.querySelector('.select-option--focus')
-    list.firstElementChild.classList.add('select-option--focus')
-    focusedItem.classList.remove('select-option--focus')
-  }
-  if (position === 'last') {
-    const focusedItem = list.querySelector('.select-option--focus')
-    list.lastElementChild.classList.add('select-option--focus')
-    focusedItem.classList.remove('select-option--focus')
-  }
-
-  setTimeout(function() {
-    list.setAttribute('aria-activedescendant', list.querySelector('.select-option--focus').id)
-
-    //For webkit (announces selected on all options though...)
-    //list.querySelector('[aria-selected="true"]').setAttribute('aria-selected','false');
-    //list.querySelector('.select-option--focus').setAttribute('aria-selected','false');
-
-    //Set focus on option.. works better for webkit
-    list.querySelector('.select-option--focus').focus()
-  }, 0)
-}
-
 const initialized = new WeakSet()
 
-/**
- * @typedef {object} SelectOptions
- * @property {string} [allLabel] Label for "All" option
- * @property {boolean} [groupToggle] Enable group toggling
- * @property {string | RegExp} [smallPattern] RegExp pattern for small text
- */
+// For backward compatibility
+export function select(element, opts = element.dataset) {
+  return new Select(element, opts)
+}
 
-/**
- * Decorate select element with custom dropdown
- * @param {Node} element Element to be decorated
- * @param {SelectOptions} opts Options
- */
-export function select(el, opts = el.dataset) {
-  const config = getConfig(opts)
-  const select = el.querySelector('.select-options')
-  const label = el.querySelector('.select-label')
-  const id = uid()
-  let selectOpen = false
-  let preventClose = false
-  let clickOutside
-  let hoverFocus
+export class Select {
+  constructor(element, opts = element.dataset) {
+    this.config = getConfig(opts)
+    this.select = element.querySelector('.select-options')
+    this.label = element.querySelector('.select-label')
+    this.id = uid()
+    this.selectOpen = false
+    this.preventClose = false
+    this.clickOutside = undefined
+    this.hoverFocus = undefined
+    this.multiselectable = this.select.multiple
 
-  function toggleSelect() {
-    selectOpen ? closeSelect() : openSelect()
-  }
+    // Search
+    this.keyString = ''
+    this.typingTimer = undefined
 
-  // Don't bother with disabled or missing elements
-  if (!label || !select || select.disabled || initialized.has(el)) {
-    return
-  }
+    this._menuItems = new WeakMap()
+    this._groups = new WeakMap()
+    this._selectAll = new Ref()
+    this._status = new Ref()
 
-  const initiallySelected = [...select.selectedOptions]
-  const placeholder = initiallySelected.find(isPlaceholder)
-  const initiallyValid = initiallySelected.filter((child) => !isPlaceholder(child))
-  setValidState(select, initiallyValid.length)
+    if (!this.label || !this.select || this.select.disabled || initialized.has(element)) {
+      return
+    }
 
-  // Set up internal lookup mechanisms
-  const menuItems = new WeakMap()
-  const groups = new WeakMap()
-  const selectAll = new Ref()
-  const status = new Ref()
+    this.initiallySelected = [...this.select.selectedOptions]
+    this.placeholder = this.select.querySelector('option[value=""][disabled]')
+    this.initiallyValid = this.initiallySelected.filter((child) => !isPlaceholder(child))
+    setValidState(this.select, this.initiallyValid.length)
 
-  function selectOption(target) {
-    const items = list.querySelectorAll('.select-option')
+    this.toggle = this.createToggle()
+    this.list = this.createOptionList()
 
-    if (!select.multiple) {
-      for (const item of items) {
-        item.classList.remove('select-option--selected')
-        item.setAttribute('aria-selected', 'false')
-        item._selectElement.selected = false
+    this.selectableItems = [...this.list.querySelectorAll('.select-option')]
+    this.searchableItems = this.selectableItems.filter((item) => {
+      return !item.classList.contains('toggle')
+    })
+    // Update status and labels on change
+    this.select.addEventListener('change', (event) => {
+      this.onChangeHandler(event)
+    })
+
+    // Update DOM
+    element.classList.add('initialized')
+    this.label.id = `dropdown-label-${this.id}`
+    this.select.after(this.toggle, this.list)
+    // Cache element to prevent double initialization
+    initialized.add(element)
+
+    const callback = (mutationList) => {
+      for (const mutation of mutationList) {
+        if (mutation.type === 'childList' || mutation.type === 'characterData') {
+          const frag = document.createDocumentFragment()
+          frag.append(...this.createOptions().filter((item) => item))
+          this.list.innerHTML = ''
+          this.list.append(frag)
+          this.placeholder = this.select.querySelector('option[value=""][disabled]')
+          this.selectableItems = [...this.list.querySelectorAll('.select-option')]
+          this.searchableItems = this.selectableItems.filter((item) => {
+            return !item.classList.contains('toggle')
+          })
+
+          this.select.dispatchEvent(new window.Event('change'))
+        }
       }
     }
 
-    target.classList.add('select-option--selected')
-    target.setAttribute('aria-selected', 'true')
-    target._selectElement.selected = select.multiple ? !target._selectElement.selected : true
+    // Create an observer instance linked to the callback function
+    const observer = new window.MutationObserver(callback)
 
-
-    if (!select.multiple) {
-      closeSelect()
-    }
-
-    select.dispatchEvent(new window.Event('change', { bubbles: true }))
+    // Start observing the target node for configured mutations
+    observer.observe(this.select, {
+      attributes: false,
+      childList: true,
+      subtree: true,
+      characterData: true
+    })
   }
 
-  function closeSelect() {
-    list.setAttribute('hidden', true)
-    list.classList.remove('select-list--top')
-    toggle.setAttribute('aria-expanded', 'false')
-    const items = list.querySelectorAll('.select-option')
-    for (const item of items) {
-      item.classList.remove('select-option--focus')
+  resolveSibling(item, position) {
+    const itemCount = this.selectableItems.length - 1
+    const itemIndex = this.selectableItems.findIndex((i) => i === item)
+
+    if (position === 'next') {
+      if (itemIndex === itemCount) {
+        return this.selectableItems[0]
+      } else {
+        return this.selectableItems[itemIndex + 1]
+      }
     }
-    selectOpen = false
-    document.removeEventListener('click', clickOutside)
-    document.removeEventListener('mouseover', hoverFocus)
-    setTimeout(function() {
-      toggle.setAttribute('tabindex', '0')
+
+    if (position === 'prev') {
+      if (itemIndex === 0) {
+        return this.selectableItems[itemCount]
+      } else {
+        return this.selectableItems[itemIndex - 1]
+      }
+    }
+  }
+
+  moveFocusTo(position) {
+    const { list } = this
+    const focusedItem = list.querySelector('.select-option--focus')
+
+    if (typeof position === 'object') {
+      if (!position.classList.contains('select-option--focus')) {
+        focusedItem.classList.remove('select-option--focus')
+        position.classList.add('select-option--focus')
+      }
+    }
+    if (position === 'selected') {
+      const selectedItem = list.querySelector('.select-option--selected') || list.querySelector('.select-option')
+      selectedItem.classList.add('select-option--focus')
+    }
+
+    if (position === 'next' || position === 'prev') {
+      const newFocusItem = this.resolveSibling(focusedItem, position)
+
+      focusedItem.classList.remove('select-option--focus')
+      newFocusItem.classList.add('select-option--focus')
+    }
+
+    setTimeout(() => {
+      list.setAttribute('aria-activedescendant', list.querySelector('.select-option--focus').id)
+
+      // Set focus on option.. works better for webkit
+      this.list.querySelector('.select-option--focus').focus({ visible: true })
     }, 0)
   }
 
-  function openSelect(e) {
-    toggle.setAttribute('tabindex', '-1')
-    toggle.setAttribute('aria-expanded', 'true')
-    list.removeAttribute('hidden')
-    moveFocusTo('selected', list)
-    selectOpen = true
+  onChangeHandler() {
+    this._selectAll.current?.classList.toggle(
+      'select-option--selected',
+      allAreSelected(this.select)
+    )
 
-    if (window.innerHeight - toggle.getBoundingClientRect().bottom < toggle.scrollHeight) {
-      list.classList.add('select-list--top')
+    const states = new Map()
+
+    // Update proxy buttons
+    for (const option of this.select.options) {
+      // if (option.disabled) continue
+      const ref = this._menuItems.get(option)
+      ref?.current.classList.toggle('select-option--selected', option.selected)
+      ref?.current.setAttribute('aria-selected', option.selected)
+      // Calculate group status
+      if (this._groups.has(option)) {
+        const button = this._groups.get(option)
+        let current = states.get(button)
+
+        if (option.selected) {
+          if (!current) current = 'selected'
+          else if (current === 'unselected') current = 'indeterminate'
+        } else if (current === 'selected') {
+          current = 'indeterminate'
+        } else if (!current) {
+          current = 'unselected'
+        }
+
+        states.set(button, current)
+      }
     }
 
-    document.addEventListener(
-      'click',
-      (clickOutside = (e) => {
-        if (e.target !== toggle && e.target !== list && !list.contains(e.target)) {
-          closeSelect()
-        }
-      })
-    )
+    // Update group status
+    for (const [button, state] of states) {
+      button.current.classList.toggle('select-option--selected', state === 'selected')
+      button.current.classList.toggle('select-option--indeterminate', state === 'indeterminate')
+    }
 
-    list.addEventListener(
-      'mouseover',
-      (hoverFocus = (e) => {
-        if (e.target.classList.contains('select-option') || e.target.closest('.select-option')) {
-          const item = e.target.classList.contains('select-option')
-            ? e.target
-            : e.target.closest('.select-option')
-          moveFocusTo(item, list)
-        }
-      })
-    )
+    const selected = [...this.select.selectedOptions].filter((option) => !isPlaceholder(option))
+    const { current: statusNode } = this._status
+
+    setValidState(this.select, selected.length)
+
+    // Update status counter
+    statusNode.dataset.count = selected.length || ''
+
+    // Update status label
+    const items = selected.length
+      ? selected
+      : this.placeholder
+        ? [this.placeholder]
+        : []
+
+    const [first, ...rest] = statusNode.childNodes
+    const next = items.length
+      ? concatLabels(items, this.config.smallPattern)
+      : ['']
+
+    for (const child of rest) child.remove()
+    first.replaceWith(...next)
   }
 
-  // The button which open the custom select list
-  const toggle = h(
-    'button',
-    {
-      type: 'button',
-      class: 'select-toggle',
-      id: `dropdown-select-${id}`,
-      'aria-expanded': 'false',
-      'aria-controls': `list-${id}`,
-      'aria-labelledby': `dropdown-label-${id} dropdown-select-${id}`,
-      'aria-haspopup': 'listbox',
-      onclick() {
-        toggleSelect()
-        setTimeout(function() {
-          if (!selectOpen) {
-            toggle.focus()
-          } else {
-            list.focus()
-          }
-        }, 0)
-      },
-      onkeydown(e) {
-        if (e.key === 'Enter' || e.key === ' ') {
-          toggleSelect()
-          setTimeout(function() {
-            if (!selectOpen) {
-              toggle.focus()
+  onClickHandler({ target }) {
+    if (
+      target.classList.contains('select-option') ||
+      target.closest('.select-option')
+    ) {
+      const item = target.classList.contains('select-option')
+        ? target
+        : target.closest('.select-option')
+
+      if (item.ariaDisabled === 'false') {
+        this.selectOption(item)
+        this.toggle.focus()
+      }
+    }
+  }
+
+  onKeydownHandler(event) {
+    if (event.key === 'Enter' || event.key === ' ') {
+      const item = this.list.querySelector('.select-option--focus')
+      if (item.ariaDisabled === 'true') {
+        event.stopPropagation()
+      } else if (item.classList.contains('toggle')) {
+        this.toggleGroup(item)
+      } else {
+        this.selectOption(item)
+        if (!this.multiselectable) {
+          this.toggle.focus()
+        }
+      }
+
+      event.preventDefault()
+    }
+
+    if (event.key === 'Escape') {
+      this.closeSelect()
+      this.toggle.focus()
+      event.preventDefault()
+    }
+    if (event.key === 'Tab') {
+      this.closeSelect()
+    }
+    if (event.key === 'ArrowDown') {
+      this.moveFocusTo('next')
+      event.preventDefault()
+    }
+    if (event.key === 'ArrowUp') {
+      this.moveFocusTo('prev')
+      event.preventDefault()
+    }
+    if (event.key === 'Home') {
+      this.moveFocusTo('first')
+      event.preventDefault()
+    }
+    if (event.key === 'End') {
+      this.moveFocusTo('last')
+      event.preventDefault()
+    }
+
+    // Search functionality
+    if (/^.$/u.test(event.key)) {
+      this.keyString += event.key
+      clearTimeout(this.typingTimer)
+      if (this.keyString) {
+        const matchedItems = this.search()
+        this.moveFocusTo(matchedItems[0])
+        this.typingTimer = setTimeout(() => {
+          this.keyString = ''
+        }, 250)
+      }
+
+      event.stopPropagation()
+    }
+  }
+
+  search() {
+    return this.searchableItems.filter((item) => {
+      const regex = new RegExp('^' + this.keyString, 'gi')
+      return item.textContent.match(regex)
+    })
+  }
+
+  toggleGroup(item) {
+    const group = item.closest('.select-group')
+
+    const expanded = item.getAttribute('aria-expanded') === 'true'
+    const options = group.querySelectorAll('.select-list > .select-option')
+
+    if (expanded) {
+      group.classList.remove('show')
+      options.forEach((option) => option.setAttribute('aria-disabled', true))
+    } else {
+      group.classList.add('show')
+      options.forEach((option) => option.setAttribute('aria-disabled', false))
+    }
+
+    this.moveFocusTo(item)
+    this.selectableItems = [...this.list.querySelectorAll('.select-group:not(.show) .select-option.toggle, .select-group.show .select-option')]
+
+    item.setAttribute('aria-expanded', !expanded)
+  }
+
+  createToggle() {
+    const { id } = this
+    const toggle = h(
+      'button',
+      {
+        type: 'button',
+        class: 'select-toggle',
+        id: `dropdown-select-${id}`,
+        'aria-expanded': 'false',
+        'aria-controls': `list-${id}`,
+        'aria-labelledby': `dropdown-label-${id} dropdown-select-${id}`,
+        'aria-haspopup': 'listbox',
+        onclick: () => {
+          this.toggleSelect()
+          setTimeout(() => {
+            if (!this.selectOpen) {
+              this.toggle.focus()
             } else {
-              list.focus()
+              this.list.focus()
             }
           }, 0)
-          e.preventDefault()
-        }
-
-        if (e.key === 'Escape') {
-          closeSelect()
-          toggle.focus()
-          e.preventDefault()
-        }
-        if (e.key === 'ArrowDown') {
-          openSelect()
-          list.focus()
-          e.preventDefault()
-        }
-      }
-    },
-    [
-      h(
-        'span',
-        {
-          'data-count': initiallyValid.length || '',
-          class: 'select-status',
-          role: 'status',
-          ref: status
         },
-        initiallyValid.length
-          ? concatLabels(initiallyValid, config.smallPattern)
-          : placeholder
-            ? resolveLabel(placeholder.label, config.smallPattern)
-            : ''
-      )
-    ]
-  )
+        onkeydown: (e) => {
+          if (e.key === 'Enter' || e.key === ' ') {
+            this.toggleSelect()
+            setTimeout(() => {
+              if (!this.selectOpen) {
+                this.toggle.focus()
+              } else {
+                this.list.focus()
+              }
+            }, 0)
+            e.preventDefault()
+          }
 
-  // The list of options
-  const list = h(
-    'ul',
-    {
-      id: `list-${id}`,
-      role: 'listbox',
-      tabindex: '-1',
-      class: 'select-list',
-      'aria-labelledby': `dropdown-label-${id}`,
-      'aria-multiselectable': select.multiple,
-      hidden: true,
-      // We want this later.
-      // 'aria-activedescendant': select.selectedOptions[0].id,
-      onclick() {
-        // Disable automatic closing on click
-        preventClose = preventClose || select.multiple
-      }
-    },
-    [
+          if (e.key === 'Escape') {
+            this.closeSelect()
+            this.toggle.focus()
+            e.preventDefault()
+          }
+          if (e.key === 'ArrowDown') {
+            this.openSelect()
+            this.list.focus()
+            e.preventDefault()
+          }
+        }
+      },
+      [
+        h(
+          'span',
+          {
+            'data-count': this.initiallyValid.length || '',
+            class: 'select-status',
+            role: 'status',
+            ref: this._status
+          },
+          this.initiallyValid.length
+            ? concatLabels(this.initiallyValid, this.config.smallPattern)
+            : this.placeholder
+              ? resolveLabel(this.placeholder.label, this.config.smallPattern)
+              : ''
+        )
+      ]
+    )
+
+    return toggle
+  }
+
+  toggleAll() {
+    const selected = !allAreSelected(this.select)
+    for (const option of this.select.options) {
+      if (!option.disabled) option.selected = selected
+    }
+    this.select.dispatchEvent(new window.Event('change'))
+  }
+
+  createOptionList() {
+    const { id, select } = this
+    const list = h(
+      'ul',
+      {
+        id: `list-${id}`,
+        role: 'listbox',
+        tabindex: '-1',
+        class: 'select-list',
+        'aria-labelledby': `dropdown-label-${id}`,
+        'aria-multiselectable': select.multiple,
+        hidden: true,
+        // We want this later.
+        // 'aria-activedescendant': select.selectedOptions[0].id,
+        onclick: () => {
+          // Disable automatic closing on click
+          this.preventClose = this.preventClose || select.multiple
+        }
+      },
+      this.createOptions()
+    )
+
+    list.addEventListener('keydown', (event) => this.onKeydownHandler(event))
+    list.addEventListener('click', (event) => this.onClickHandler(event))
+
+    return list
+  }
+
+  createOptions() {
+    const { id, select, config, _selectAll, _menuItems, _groups } = this
+
+    return [
       // Toggle all button
       select.multiple && config.allLabel
         ? h(
-          'button',
+          'li',
           {
-            ref: selectAll,
-            type: 'button',
-            class: 'select-option multiple dropdown-item',
-            onclick(event) {
-              const selected = !allAreSelected(select)
-              for (const option of select.options) {
-                if (!option.disabled) option.selected = selected
-              }
-              select.dispatchEvent(new window.Event('change'))
+            ref: _selectAll,
+            role: 'option',
+            id: `select-all-option-${id}`,
+            class: 'select-option multiple',
+            'aria-selected': !allAreSelected(select),
+            tabindex: '-1',
+            onclick: (event) => {
+              this.toggleAll()
               event.preventDefault()
+            },
+            onkeydown: (event) => {
+              if (event.key === 'Enter' || event.key === ' ') {
+                this.toggleAll()
+                event.preventDefault()
+              }
             }
           },
           config.allLabel
@@ -287,9 +447,8 @@ export function select(el, opts = el.dataset) {
             const group = h(
               'fieldset',
               {
-                'data-bs-target': `#list-${id}`,
                 id: `group-${index}-list-${id}`,
-                class: 'select-list dropdown-menu',
+                class: 'select-list',
                 'aria-labelledby': `group-${index}-trigger-${id}`
               },
               children
@@ -297,7 +456,7 @@ export function select(el, opts = el.dataset) {
 
             // Add to internal lookup table)
             for (const option of child.children) {
-              groups.set(option, button)
+              _groups.set(option, button)
             }
 
             // Wrap children
@@ -317,33 +476,21 @@ export function select(el, opts = el.dataset) {
                   {
                     role: 'option',
                     ref: button,
-                    // type: 'button',
                     id: `group-${index}-trigger-${id}`,
                     class: `select-option ${select.multiple ? 'multiple' : ''
-                      } toggle dropdown-item`,
-                    // 'aria-expanded': 'true',
-                    // 'aria-controls': `group-${index}-list-${id}`,
+                      } toggle`,
+                    'aria-expanded': 'true',
+                    tabindex: '-1',
+                    'aria-controls': `group-${index}-list-${id}`,
 
-                    // TODO: NEED to add keyboard event to handle enter or spacebar to select option.
-                    onclick() {
+                    onclick(event) {
                       expanded = !expanded
+
                       this.setAttribute('aria-expanded', expanded ? 'true' : 'false')
                       wrapper.current.classList.toggle('show', expanded)
 
                       // Disable automatic closing on click
-                      preventClose = true
-
-                      // Ugly hack to prevent Bootstrap dropdown from including
-                      // hidden items in key navigation
-                      for (const option of child.children) {
-                        const button = menuItems.get(option)
-                        button.current.classList.toggle('disabled', !expanded)
-                      }
-
-                      // Update dropdown position
-                      // dropdown.update()
-                      //
-                      // TODO: we need to update the position maybe
+                      this.preventClose = true
                     }
                   },
                   label
@@ -362,7 +509,7 @@ export function select(el, opts = el.dataset) {
 
         // Add to internal element reverse lookup table
         const button = new Ref()
-        menuItems.set(child, button)
+        _menuItems.set(child, button)
 
         const listItem = h(
           'li',
@@ -385,146 +532,93 @@ export function select(el, opts = el.dataset) {
         return listItem
       })
     ]
-  )
+  }
 
-  list.addEventListener('keydown', function(event) {
-    if (event.key === 'Enter' || event.key === ' ') {
-      const item = list.querySelector('.select-option--focus')
+  toggleSelect() {
+    this.selectOpen ? this.closeSelect() : this.openSelect()
+  }
 
-      if (item.ariaDisabled === 'true') {
-        event.stopPropagation()
-      } else {
-        selectOption(item)
-        if (!select.multiple) {
-          toggle.focus()
+  closeSelect() {
+    this.list.setAttribute('hidden', true)
+    this.list.classList.remove('select-list--top')
+    this.toggle.setAttribute('aria-expanded', 'false')
+
+    const items = this.list.querySelectorAll('.select-option')
+    for (const item of items) {
+      item.classList.remove('select-option--focus')
+    }
+
+    this.selectOpen = false
+
+    document.removeEventListener('click', this.onClickOutsideHandler)
+    document.removeEventListener('mouseover', this.onMouseoverHandler)
+
+    setTimeout(() => {
+      this.toggle.setAttribute('tabindex', '0')
+    }, 0)
+  }
+
+  openSelect() {
+    const { toggle, list } = this
+
+    this.toggle.setAttribute('tabindex', '-1')
+    this.toggle.setAttribute('aria-expanded', 'true')
+    this.list.removeAttribute('hidden')
+    this.moveFocusTo('selected')
+    this.selectOpen = true
+
+    const listBottomPosition = list.getBoundingClientRect().bottom
+    if (window.innerHeight - listBottomPosition < toggle.scrollHeight) {
+      list.classList.add('select-list--top')
+    }
+
+    document.addEventListener(
+      'click',
+      (this.clickOutside = ({ target }) => {
+        if (target !== this.toggle && target !== this.list && !this.list.contains(target)) {
+          this.closeSelect()
+        }
+      })
+    )
+
+    this.list.addEventListener(
+      'mouseover',
+      (this.hoverFocus = ({ target }) => {
+        const item = target.closest('.select-option')
+        if (item) {
+          this.moveFocusTo(item)
+        }
+      })
+    )
+  }
+
+  selectOption(target) {
+    const items = this.list.querySelectorAll('.select-option')
+
+    if (!this.multiselectable) {
+      for (const item of items) {
+        item.classList.remove('select-option--selected')
+        item.setAttribute('aria-selected', 'false')
+        if (item.selectElement) {
+          item._selectElement.selected = false
         }
       }
-
-      event.preventDefault()
-    }
-    if (event.key === 'Escape') {
-      closeSelect()
-      toggle.focus()
-      event.preventDefault()
-    }
-    if (event.key === 'Tab') {
-      closeSelect()
-    }
-    if (event.key === 'ArrowDown') {
-      moveFocusTo('next', list)
-      event.preventDefault()
-    }
-    if (event.key === 'ArrowUp') {
-      moveFocusTo('prev', list)
-      event.preventDefault()
-    }
-    if (event.key === 'Home') {
-      moveFocusTo('first', list)
-      event.preventDefault()
-    }
-    if (event.key === 'End') {
-      moveFocusTo('last', list)
-      event.preventDefault()
     }
 
-    // Search functionality
-    // if (/^.$/u.test(e.key)) {
-    //   keyString += e.key
-    //   clearTimeout(typingTimer)
-    //   if (keyString) {
-    //     let matchedItems = filterSelect(keyString, options)
-    //     moveFocusTo(matchedItems[0])
-    //     typingTimer = setTimeout(function () {
-    //       keyString = ''
-    //     }, 250)
-    //   }
-    // }
-  })
+    target.classList.add('select-option--selected')
+    target.setAttribute('aria-selected', 'true')
 
-  list.addEventListener('click', function(event) {
-    if (
-      event.target.classList.contains('select-option') ||
-      event.target.closest('.select-option')
-    ) {
-      const item = event.target.classList.contains('select-option')
-        ? event.target
-        : event.target.closest('.select-option')
-
-      if (item.ariaDisabled === 'false') {
-        selectOption(item)
-        toggle.focus()
-      }
-    }
-  })
-
-  // Prevent automatic close on click
-  el.addEventListener('hide.bs.dropdown', function(event) {
-    if (preventClose) {
-      preventClose = false
-      event.preventDefault()
-    }
-  })
-
-  // Update status and labels on change
-  select.addEventListener('change', function() {
-    selectAll.current?.classList.toggle('select-option--selected', allAreSelected(select))
-    const states = new Map()
-
-    // Update proxy buttons
-    for (const option of select.options) {
-      // if (option.disabled) continue
-      const ref = menuItems.get(option)
-      ref?.current.classList.toggle('select-option--selected', option.selected)
-      ref?.current.setAttribute('aria-selected', option.selected)
-      // Calculate group status
-      if (groups.has(option)) {
-        const button = groups.get(option)
-        let current = states.get(button)
-
-        if (option.selected) {
-          if (!current) current = 'selected'
-          else if (current === 'unselected') current = 'indeterminate'
-        } else if (current === 'selected') {
-          current = 'indeterminate'
-        } else if (!current) {
-          current = 'unselected'
-        }
-
-        states.set(button, current)
-      }
+    if (target._selectElement) {
+      target._selectElement.selected = this.multiselectable ? !target._selectElement.selected : true
     }
 
-    // Update group status
-    for (const [button, state] of states) {
-      button.current.classList.toggle('select-option--selected', state === 'selected')
-      button.current.classList.toggle('select-option--indeterminate', state === 'indeterminate')
+
+    if (!this.multiselectable) {
+      this.closeSelect()
     }
 
-    const selected = [...select.selectedOptions].filter((option) => !isPlaceholder(option))
-    const { current: statusNode } = status
-
-    setValidState(select, selected.length)
-
-    // Update status counter
-    statusNode.dataset.count = selected.length || ''
-
-    // Update status label
-    const items = selected.length ? selected : placeholder ? [placeholder] : []
-    const [first, ...rest] = statusNode.childNodes
-
-    const next = items.length ? concatLabels(items, config.smallPattern) : ['']
-
-    for (const child of rest) child.remove()
-    first.replaceWith(...next)
-  })
-
-  // Update DOM
-  el.classList.add('initialized')
-  label.id = `dropdown-label-${id}`
-  select.after(toggle, list)
-
-  // Cache element to prevent double initialization
-  initialized.add(el)
+    this.select.dispatchEvent(new window.Event('change', { bubbles: true }))
+  }
 }
 
 /**
@@ -562,8 +656,8 @@ function setValidState(element, valid) {
  * @returns {boolean}
  */
 function isPlaceholder(option) {
-  const { value, disabled, hidden } = option
-  return !value && disabled && hidden
+  const { value, disabled } = option
+  return !value && disabled
 }
 
 /**
